// apps/web/src/types/payload-types.ts
export interface User {
  id: number;
  name?: string | null;
  avatar?: string | null;
  bio?: string | null;
  city?: string | null;
  country?: string | null;
  likedArtworks?: (number | Artwork)[] | null;
  savedEvents?: (number | Event)[] | null;
  followedArtists?: (number | Artist)[] | null;
  followedGalleries?: (number | Gallery)[] | null;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
  | {
    id: string;
    createdAt?: string | null;
    expiresAt: string;
  }[]
  | null;
  password?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "artworks".
 */
export interface Artwork {
  id: number;
  title: string;
  artist: number | Artist;
  gallery?: (number | null) | Gallery;
  year?: number | null;
  image?: (number | null) | Media;
  medium?: string | null;
  dimensions?: string | null;
  category?: string | null;
  description?: string | null;
  likes?: number | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "artists".
 */
export interface Artist {
  id: number;
  name: string;
  country?: string | null;
  birthYear?: number | null;
  image?: (number | null) | Media;
  bio?: string | null;
  disciplines?:
  | {
    name?: string | null;
    id?: string | null;
  }[]
  | null;
  movements?:
  | {
    name?: string | null;
    id?: string | null;
  }[]
  | null;
  type?: ('Individual' | 'Collective') | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "galleries".
 */
export interface Gallery {
  id: number;
  name: string;
  location?: string | null;
  image?: (number | null) | Media;
  description?: string | null;
  type?: ('Gallery' | 'Museum') | null;
  institutionType?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "events".
 */
export interface Event {
  id: number;
  name: string;
  location?: string | null;
  startDate?: string | null;
  endDate?: string | null;
  image?: (number | null) | Media;
  type?: string | null;
  rawDate?: string | null;
  description?: string | null;
  gallery?: (number | null) | Gallery;
  updatedAt: string;
  createdAt: string;
}